+++ lib/src/bullet_train_client.dart
@@ -1,3 +1,6 @@
+import 'dart:io';
+import 'package:rxdart/subjects.dart';
+
 import 'store/storage_provider.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
@@ -13,9 +16,13 @@ import 'store/storage/persistant_store.dart';
 ///
 /// [config] configuration for http client and endpoints
 /// [apiKey] api key for your enviornment
+/// [seeds] default values before update from bullet train env.
+///
+
 class BulletTrainClient {
   static final String authHeader = 'X-Environment-Key';
   static final String acceptHeader = 'Accept';
+  static final String userAgentHeader = 'User-Agent';
 
   final String apiKey;
   final BulletTrainConfig config;
@@ -26,6 +33,7 @@ class BulletTrainClient {
       @required this.apiKey,
       List<Flag> seeds})
       : assert(apiKey != null, 'Missing Bullet-train.io apiKey') {
+    bulletTrainDebug = config.isDebug;
     switch (config.storeType) {
       case StoreType.persistant:
         storage = StorageProvider(PersistantStore(), password: config.password);
@@ -44,9 +52,19 @@ class BulletTrainClient {
   }
 
   /// Simple implementation of Http Client
-  Dio get _api => Dio(config.clientOptions)
-    ..options.headers[authHeader] = apiKey
-    ..options.headers[acceptHeader] = 'application/json';
+  Dio _api() {
+    var dio = Dio(config.clientOptions)
+      ..options.headers[userAgentHeader] =
+          'BulletTrainFlutterSDK(${Platform.operatingSystem}/${Platform.version})'
+      ..options.headers[authHeader] = apiKey
+      ..options.headers[acceptHeader] = 'application/json';
+    if (config.isDebug) {
+      dio.interceptors.add(config.isDebug
+          ? LogInterceptor(requestHeader: false, responseBody: true)
+          : null);
+    }
+    return dio;
+  }
 
   /// Get a list of existing Features for the given environment and user
   ///
@@ -54,72 +72,40 @@ class BulletTrainClient {
   /// Returns a list of feature flags
   Future<List<Flag>> getFeatureFlags(
       {FeatureUser user, bool reload = true}) async {
+    if (!reload) {
+      var result = await storage.getAll();
+      result.sort((a, b) => a.feature.name.compareTo(b.feature.name));
+      return result;
+    }
+    return user == null ? await _getFlags() : await _getUserFlags(user);
+  }
+
+  /// Check if Feature flag exist and is enabled
+  ///
+  /// [featureName] an identifier for the feature
+  /// [user] an identifier for the user
+  /// Returns true if feature flag exist and enabled, false otherwise
+  Future<bool> hasFeatureFlag(String featureName, {FeatureUser user}) async {
     try {
-      if (user == null) {
-        if (!reload) {
-          return await storage.getAll();
-        }
-
-        var params = <String, dynamic>{'page': '1'};
-        var response = await _api.get<List<dynamic>>(config.flagsURI,
-            queryParameters: params);
-        var returnList = <Flag>[];
-        if (response.statusCode == 200) {
-          var list = response.data
-              .map<Flag>((dynamic e) => Flag.fromMap(e as Map<String, dynamic>))
-              .toList();
-
-          await storage.saveAll(list);
-
-          returnList = list;
-        } else {
-          returnList = await storage.getAll();
-        }
-        return returnList;
-      } else {
-        if (!reload) {
-          return await storage.getAll();
-        }
-        var response = await _api
-            .get<List<dynamic>>('${config.flagsURI}${'/${user.identifier}'}');
-        var returnList = <Flag>[];
-        if (response.statusCode == 200) {
-          var list = response.data
-              .map<Flag>((dynamic e) => Flag.fromMap(e as Map<String, dynamic>))
-              .toList();
-          await storage.saveAll(list);
-          returnList = list;
-        } else {
-          returnList = await storage.getAll();
-        }
-        return returnList;
-      }
-    } on DioError catch (_) {
-      throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
-    } on FormatException catch (_) {
-      throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
+      var features = await getFeatureFlags(user: user, reload: false);
+      var feature = features.firstWhere(
+          (element) =>
+              element.feature.name == featureName && element.enabled == true,
+          orElse: () => null);
+      return feature != null;
     } on Exception catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.genericError);
     }
   }
 
-  /// Check if Feature flag exist and is enabled
-  ///
-  /// [featureId] an identifier for the feature
-  /// [user] an identifier for the user
-  /// Returns true if feature flag exist and enabled, false otherwise
-  Future<bool> hasFeatureFlag(String featureId, {FeatureUser user}) async {
+  Future<bool> isFeatureFlagEnabled(String featureName,
+      {FeatureUser user}) async {
     try {
       var features = await getFeatureFlags(user: user, reload: false);
       var feature = features.firstWhere(
-          (element) =>
-              element.feature.name == featureId && element.enabled == true,
+          (element) => element.feature.name == featureName,
           orElse: () => null);
-      return feature != null;
-    } on DioError catch (_) {
-      throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
-    } on FormatException catch (_) {
-      throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
+      return feature?.enabled;
     } on Exception catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.genericError);
     }
@@ -136,7 +122,8 @@ class BulletTrainClient {
               element.feature.name == featureId && element.enabled == true,
           orElse: () => null);
       return feature?.stateValue;
-    } on DioError catch (_) {
+    } on DioError catch (e) {
+      log('getFeatureFlagValue dioError: ${e.toString()}');
       throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
     } on FormatException catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
@@ -151,7 +138,9 @@ class BulletTrainClient {
       var result = await _getUserTraits(user);
       return result.firstWhere((element) => element.key == key,
           orElse: () => null);
-    } on DioError catch (_) {
+    } on DioError catch (e) {
+      log('getTrait dioError: ${e?.error}');
+
       throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
     } on FormatException catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
@@ -164,11 +153,62 @@ class BulletTrainClient {
   Future<List<Trait>> getTraits(FeatureUser user, {List<String> keys}) async {
     try {
       var result = await _getUserTraits(user);
-      if (keys == null || keys.isEmpty) {
+      if (keys == null) {
         return result;
       }
+      var filtered =
+          result.where((element) => keys.contains(element.key)).toList();
       return result.where((element) => keys.contains(element.key)).toList();
-    } on DioError catch (_) {
+    } on DioError catch (e) {
+      log('getTraits dioError: ${e?.error}');
+      throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
+    } on FormatException catch (_) {
+      throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
+    } on Exception catch (_) {
+      throw BulletTrainException(BulletTrainExceptionType.genericError);
+    }
+  }
+
+  Future<List<Flag>> _getFlags() async {
+    try {
+      var response = await _api().get<List<dynamic>>(config.flagsURI);
+      if (response.statusCode == 200) {
+        var list = response.data
+            .map<Flag>((dynamic e) => Flag.fromMap(e as Map<String, dynamic>))
+            .toList();
+        await storage.saveAll(list);
+        list.sort((a, b) => a.feature.name.compareTo(b.feature.name));
+        return list;
+      }
+      return [];
+    } on DioError catch (e) {
+      log('_getFlags dioError: ${e?.error}');
+      throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
+    } on FormatException catch (_) {
+      throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
+    } on Exception catch (_) {
+      throw BulletTrainException(BulletTrainExceptionType.genericError);
+    }
+  }
+
+  // Internal list of [user] flags
+  Future<List<Flag>> _getUserFlags(FeatureUser user) async {
+    try {
+      var params = {'identifier': user.identifier};
+      var response = await _api().get<Map<String, dynamic>>(
+          '${config.identitiesURI}',
+          queryParameters: params);
+
+      if (response.statusCode == 200) {
+        var data = FlagAndTraits.fromMap(response.data)?.flags ?? [];
+        data.sort((a, b) => a.feature.name.compareTo(b.feature.name));
+        await storage.saveAll(data);
+
+        return data;
+      }
+      return [];
+    } on DioError catch (e) {
+      log('getUserTraits dioError: ${e?.error}');
       throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
     } on FormatException catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
@@ -181,15 +221,17 @@ class BulletTrainClient {
   Future<List<Trait>> _getUserTraits(FeatureUser user) async {
     try {
       var params = {'identifier': user.identifier};
-      var response = await _api.get<dynamic>('${config.identitiesURI}',
+      var response = await _api().get<Map<String, dynamic>>(
+          '${config.identitiesURI}',
           queryParameters: params);
 
       if (response.statusCode == 200) {
-        var data = FlagAndTraits.fromMap(response.data as Map<String, dynamic>);
-        return data.traits ?? [];
+        return FlagAndTraits.fromMap(response.data)?.traits ?? [];
       }
       return [];
-    } on DioError catch (_) {
+    } on DioError catch (e) {
+      log('getUserTraits dioError: ${e?.error}');
+
       throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
     } on FormatException catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
@@ -208,9 +250,10 @@ class BulletTrainClient {
     try {
       var trait = toUpdate.copyWith(identity: user);
       var response =
-          await _api.post<dynamic>(config.traitsURI, data: trait.toJson());
+          await _api().post<dynamic>(config.traitsURI, data: trait.toJson());
       return Trait.fromMap(response.data as Map<String, dynamic>);
-    } on DioError catch (_) {
+    } on DioError catch (e) {
+      log('_postUserTraits dioError: ${e?.error}');
       throw BulletTrainException(BulletTrainExceptionType.connectionSettings);
     } on FormatException catch (_) {
       throw BulletTrainException(BulletTrainExceptionType.wrongFlagFormat);
@@ -219,5 +262,17 @@ class BulletTrainClient {
     }
   }
 
+  /// clear all data from storage
   Future<bool> clearStore() async => storage.clear();
+
+  /// stream for listener
+  Stream<Flag> stream(String key) => storage.stream(key);
+
+  /// basic stream
+  BehaviorSubject<Flag> subject(String key) => storage.subject(key);
+
+  /// test toggle feature
+  ///
+  Future<bool> testToggle(String featureName) =>
+      storage.togggleFeature(featureName);
 }