+++ lib/src/model/index.dart
@@ -3,5 +3,4 @@ export 'feature_user.dart';
 export 'flag.dart';
 export 'flag_and_traits.dart';
 export 'flag_type.dart';
-export 'flag_type_x.dart';
 export 'trait.dart';